requirements.txt
Django==5.0.1
django-mptt==0.16.0
django-seed==0.3.1
psycopg2-binary==2.9.9
Pillow==10.2.0
environs==10.3.0
django-bootstrap5==23.4
